pyproject.toml
[project]
name = "hitl-seg-backend"
version = "0.1.0"
description = "HITL segmentation backend: DINOv3-sat + UperNet with SAM3 interactive labeling"
requires-python = ">=3.11"
license = {text = "CC-BY-NC-4.0"}

dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",

    # ML / Deep learning
    "torch>=2.4",
    "torchvision>=0.19",
    "transformers>=4.45",
    "timm>=1.0.17",

    # Geospatial
    "rasterio>=1.3",
    "geopandas>=1.0",
    "shapely>=2.0",
    "pyproj>=3.6",
    "fiona>=1.9",

    # Data / image processing
    "numpy>=1.26,<2",
    "Pillow>=10.0",
    "opencv-python-headless>=4.9",
    "albumentations>=1.4",

    # HTTP client (XYZ/WMS tile fetching)
    "httpx>=0.27",

    # Config
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",

    # Dashboard
    "gradio>=4.40",

    # Utilities
    "tqdm>=4.66",
    "structlog>=24.0",

    # SAM3 — installed separately from vendor submodule:
    #   pip install -e vendor/sam3
    "sam3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["hitl*"]

[tool.ruff]
line-length = 100
target-version = "py311"
